refactor: following list page
