+++ src/pages/profile/[id]/follows/index.page.tsx
@@ -1,43 +1,31 @@
-'use client';
-
+import { type GetServerSidePropsContext } from 'next';
 import { useFetFollowList } from '@/apis/follow';
-import FollowingList from '@/app/profile/[id]/follows/FollowingList';
-import { useGetMeId } from '@/app/profile/[id]/follows/index.hooks';
-import MyFollowerList from '@/app/profile/[id]/follows/MyFollowerList';
 import Header from '@/components/Header/Header';
 import FullTab from '@/components/Tab/FullTab';
 import { useTab } from '@/components/Tab/Tab.hooks';
 import { ROUTER } from '@/constants/router';
 import useSearchParamsTypedValue from '@/hooks/useSearchParamsTypedValue';
+import FollowingList from '@/pages/profile/[id]/follows/FollowingList';
+import { useGetMeId } from '@/pages/profile/[id]/follows/index.hooks';
+import MyFollowerList from '@/pages/profile/[id]/follows/MyFollowerList';
 import { css } from '@/styled-system/css';
 
 type FollowTabType = 'following' | 'follower';
 
-function FollowListPage({ params }: { params: { id: string } }) {
-  const { searchParams } = useSearchParamsTypedValue<FollowTabType>('tab');
-  const initTabId = searchParams ?? 'following';
+export async function getServerSideProps(context: GetServerSidePropsContext) {
+  return {
+    props: {
+      params: { id: context.query.id },
+    },
+  };
+}
 
+function FollowListPage({ params }: { params: { id: string } }) {
   const currentMemberId = Number(params.id);
 
   const { data, refetch, isLoading, isFetching } = useFetFollowList(Number(params.id));
 
-  const followingCount = data?.followingList.length ?? '';
-  const followerCount = data?.followerList.length ?? '';
-
-  const _tabs = [
-    {
-      id: 'following',
-      tabName: `팔로잉 ${followingCount}`,
-      href: ROUTER.PROFILE.FOLLOW_LIST(currentMemberId, 'following'),
-    },
-    {
-      id: 'follower',
-      tabName: `팔로워 ${followerCount}`,
-      href: ROUTER.PROFILE.FOLLOW_LIST(currentMemberId, 'follower'),
-    },
-  ];
-
-  const { tabs, activeTab, onTabClick } = useTab(_tabs, initTabId);
+  const { tabs, activeTab, onTabClick } = useFollowTab(currentMemberId);
 
   const myId = useGetMeId();
   const isMyFollowList = myId === currentMemberId;
@@ -97,3 +85,28 @@ const headerCss = css({
     textStyle: 'subtitle1',
   },
 });
+
+const useFollowTab = (memberId: number) => {
+  const { searchParams } = useSearchParamsTypedValue<FollowTabType>('tab');
+  const initTabId = searchParams ?? 'following';
+
+  const { data } = useFetFollowList(memberId);
+
+  const followingCount = data?.followingList.length ?? '';
+  const followerCount = data?.followerList.length ?? '';
+
+  const _tabs = [
+    {
+      id: 'following',
+      tabName: `팔로잉 ${followingCount}`,
+      href: ROUTER.PROFILE.FOLLOW_LIST(memberId, 'following'),
+    },
+    {
+      id: 'follower',
+      tabName: `팔로워 ${followerCount}`,
+      href: ROUTER.PROFILE.FOLLOW_LIST(memberId, 'follower'),
+    },
+  ];
+
+  return useTab(_tabs, initTabId);
+};
